Completed Part 4d exercises

+++ part4/controllers/blogs.js
@@ -1,9 +1,21 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
-const { response } = require('express')
+//const { response } = require('express')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+// const getTokenFrom = request => {
+//     const authorization = request.get('authorization')
+
+//     if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+//         return authorization.substring(7)
+//     }
+
+//     return null
+// }
 
 blogsRouter.get('/', async (req, res) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
     res.json(blogs)
     
     // Blog.find({}).then(blogs => {
@@ -15,7 +27,7 @@ blogsRouter.get('/:id', async (req, res, next) => {
     const blog = await Blog.findById(req.params.id)
 
     if (blog) {
-        response.json(blog)
+        res.json(blog)
     } else {
         res.status(404).end()
     }
@@ -32,9 +44,23 @@ blogsRouter.get('/:id', async (req, res, next) => {
 })
 
 blogsRouter.post('/', async (req, res, next) => {
+    //const token = getTokenFrom(req)
+    //const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+    // if (!decodedToken.id) {
+    //     return res.status(401).json({ error: 'token is missing or invalid' })
+    // }
+
     let newBlog = new Blog(req.body)
+    //let user = await User.findOne({ username: 'tranhuy' }).exec()
+    //let user = await User.findById(decodedToken.id)
+    let user = req.user
+    newBlog.user = user
 
     const savedBlog = await newBlog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     res.status(201).json(savedBlog)
 
     // newBlog.save()
@@ -45,12 +71,28 @@ blogsRouter.post('/', async (req, res, next) => {
 })
 
 blogsRouter.delete('/:id', async (req, res, next) => {
+    //const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    const blogToDelete = await Blog.findById(req.params.id)
+    const blogOwnerId = await blogToDelete.user.toString()
+
+    // if (!decodedToken.id) {
+    //     return res.status(401).json({ error: 'token is missing or invalid' })
+    // }
+
+    // if (blogOwnerId !== decodedToken.id) {
+    //     return res.status(403).json({ error: 'user with provided token does not have permission to delete blog' })
+    // }   
+
+    if (blogOwnerId !== req.user.id) {
+        return res.status(403).json({ error: 'user with provided token does not have permission to delete blog' })
+    }
+ 
     const deletedBlog = await Blog.findByIdAndRemove(req.params.id)
 
     if (deletedBlog) {
         res.status(204).end()
     } else {
-        res.status(404).send('Contact not found')
+        res.status(404).send('Blog not found')
     }
 })
